Create configuration services folder on common
