+++ plugins/wazuh-core/common/services/configuration/configuration-provider.ts
