+++ plugins/wazuh-core/common/services/configuration/configuration-store.md
