+++ plugins/wazuh-core/common/services/configuration/configuration.md
