+++ plugins/wazuh-core/common/services/configuration/configuration.test.ts
