+++ plugins/wazuh-core/common/services/configuration/index.ts
@@ -1,0 +1,3 @@
+export * from './configuration';
+export * from './configuration-provider';
+export * from './configuration-store';
