--- plugins/wazuh-core/server/services/configuration-repository.ts
@@ -1,4 +1,0 @@
-export interface IConfigurationRepository {
-    getConfiguration(key: string): Promise<any>;
-    setConfiguration?(key: string, value: any): Promise<void>;
-  }
